refactor a couple of functions from snake into a utils file
